🔧 merge utilities and export all functions and types
