+++ package.json
@@ -19,11 +19,11 @@
         "./utils": {
             "import": {
                 "types": "./dist/esm/utils.d.mts",
-                "default": "./dist/esm/utils/index.mjs"
+                "default": "./dist/esm/utils.mjs"
             },
             "require": {
                 "types": "./dist/cjs/utils.d.ts",
-                "default": "./dist/cjs/utils/index.js"
+                "default": "./dist/cjs/utils.js"
             }
         },
         "./package.json": "./package.json"
